+++ apps/erlangbridge/src/gen_lsp_config_server.erl
@@ -6,7 +6,8 @@
 -export([init/1, handle_call/3, handle_cast/2, handle_info/2, terminate/2, code_change/3]).
 -export([standard_modules/0, bifs/0]).
 -export([update_config/2, root/0, tmpdir/0, codeLensEnabled/0, includePaths/0, linting/0, 
-        verbose/0, autosave/0, proxy/0, search_exclude/0, formatting_line_length/0]).
+        verbose/0, autosave/0, proxy/0, search_exclude/0, formatting_line_length/0,
+        inlayHintsEnabled/0]).
 
 -define(SERVER, ?MODULE).
 
@@ -36,6 +37,9 @@ root() ->
 
 codeLensEnabled() ->
     get_config_entry(erlang, codeLensEnabled, false).
+
+inlayHintsEnabled() ->
+    get_config_entry(erlang, inlayHintsEnabled, false).
 
 includePaths() ->
     get_config_entry(erlang, includePaths, []).