+++ lib/erlangSettings.ts
@@ -5,6 +5,7 @@ export interface ErlangSettings {
 	includePaths : string[];
 	linting: boolean;
 	codeLensEnabled : boolean;
+	inlayHintsEnabled: boolean;
 	verbose: boolean;
 	debuggerRunMode : string;
 	/// workspace.rootPath, since VSCode 1.78 workspace.workspaceFolders[0].Uri.path